refactor State and Config in separate files
